move .c & .h files to lexyacc-code/
Renamed calc3b to calc3i, and added the original calc3b so that we can compare the old pseudo code with our version. Checking number of args for x86-64-driver.sh. We might want to validate that the input file has a '.calc' ending? Executing Makefile in x86-64-driver.sh.